requirements.txt
accelerate==1.1.1
datasets==3.1.0
evaluate==0.4.3
huggingface==0.0.1
huggingface-hub==0.26.2
nltk==3.9.1
pillow==11.0.0
tensorboard==2.18.0
transformers==4.49.0
torch==2.5.1
numpy==2.1.3
tqdm==4.67.1
pandas==2.2.3

